sp1 & r0 compiles with host
