+++ raiko-guests/powdr/Cargo.toml
@@ -12,7 +12,7 @@ path = "src/main.rs"
 
 
 [dependencies]
-raiko-primitives = { path = "../../primitives", features = ["revm"] }
+raiko-primitives = { path = "../../primitives" }
 raiko-lib = { path = "../../lib", features = ["taiko"] }
 powdr-riscv-runtime ={ git = "https://github.com/ceciliaz030/powdr", branch = "+nightly" }
