+++ raiko-guests/sgx/script/Cargo.toml
@@ -12,14 +12,14 @@ serde_with = {version = "3.4.0", optional = true}
 tracing = {version = "0.1", optional = true}
 tokio = { version = "^1.23", features = ["full"], optional = true}
 bincode = {version = "1.3.3", optional = true}
-zeth-lib = { path = "../../../lib", features = ["taiko", "std"] , optional = true}
-zeth-primitives = { path = "../../../primitives", features = ["taiko", "revm"] , optional = true}
+raiko-lib = { path = "../../../lib", features = ["taiko", "std"] , optional = true}
+raiko-primitives = { path = "../../../primitives", features = ["taiko"] , optional = true}
 
 
 [features]
 enable = [
-    "zeth-lib",
-    "zeth-primitives",
+    "raiko-lib",
+    "raiko-primitives",
     "serde",
     "serde_json",
     "serde_with",
