+++ raiko-guests/succinct/script/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 alloy-primitives = { version = "0.6", default-features = false, optional = true }
 serde = { version = "1.0", optional = true}
 serde_json = { version = "1.0", optional = true }
-zeth-lib = { path = "../../../lib", features = ["taiko", "std"] , optional = true }
+raiko-lib = { path = "../../../lib", features = ["taiko", "std"] , optional = true }
 sp1-core = { git = "https://github.com/succinctlabs/sp1.git" , optional = true }
 anyhow = {version = "1.0", optional = true }
 
@@ -18,7 +18,7 @@ sp1-helper = { git = "https://github.com/succinctlabs/sp1.git" , optional = true
 enable = [
     "serde",
     "serde_json",
-    "zeth-lib",
+    "raiko-lib",
     "sp1-core",
     "anyhow",
     "sp1-helper",
