+++ raiko-guests/succinct/script/src/lib.rs
@@ -4,7 +4,7 @@ use std::env;
 
 use serde::{Deserialize, Serialize};
 use sp1_core::{utils, SP1Prover, SP1Stdin, SP1Verifier};
-use zeth_lib::input::{GuestInput, GuestOutput};
+use raiko_lib::input::{GuestInput, GuestOutput};
 
 const ELF: &[u8] = include_bytes!("../../program/elf/riscv32im-succinct-zkvm-elf");
