+++ raiko-host/src/main.rs
@@ -2,8 +2,10 @@
 #![feature(generic_const_exprs)]
 #![feature(absolute_path)]
 #![feature(offset_of)]
-// TODO(Cecilia): sp1 need this due to compiler versioning, fix toolchain condition on guest type
-#![feature(generic_const_exprs)]
+
+// // TODO(Cecilia): sp1 need this due to compiler versioning, fix toolchain condition on guest type
+// #![feature(generic_const_exprs)]
+
 // Copyright 2023 RISC Zero, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
