--- raiko-host/src/prover/json_rpc.rs
@@ -1,31 +1,0 @@
-/// Common utilities for json-rpc
-use serde::Deserialize;
-use serde::Serialize;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct JsonRpcError {
-    pub code: i32,
-    pub message: String,
-}
-
-#[derive(Debug, Serialize)]
-pub struct JsonRpcResponseError {
-    pub jsonrpc: String,
-    pub id: serde_json::Value,
-    pub error: JsonRpcError,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct JsonRpcResponse<T> {
-    pub jsonrpc: String,
-    pub id: serde_json::Value,
-    pub result: Option<T>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct JsonRpcRequest<T: Serialize> {
-    pub jsonrpc: String,
-    pub id: serde_json::Value,
-    pub method: String,
-    pub params: T,
-}
