+++ raiko-host/src/prover/mod.rs
@@ -2,6 +2,5 @@ pub mod cache;
 pub mod context;
 pub mod error;
 pub mod execution;
-pub mod json_rpc;
 pub mod request;
 pub mod server;
