+++ raiko-host/src/prover/request.rs
@@ -1,5 +1,6 @@
 use core::fmt::Debug;
 
+use alloy_primitives::{Address, B256};
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr};
 
@@ -32,4 +33,32 @@ pub struct ProofRequest<P> {
     pub prover: Address,
     // Generic proof parameters which has to match with the type
     pub proof_param: P,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct JsonRpcError {
+    pub code: i32,
+    pub message: String,
+}
+
+#[derive(Debug, Serialize)]
+pub struct JsonRpcResponseError {
+    pub jsonrpc: String,
+    pub id: serde_json::Value,
+    pub error: JsonRpcError,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct JsonRpcResponse<T> {
+    pub jsonrpc: String,
+    pub id: serde_json::Value,
+    pub result: Option<T>,
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub struct JsonRpcRequest<T: Serialize> {
+    pub jsonrpc: String,
+    pub id: serde_json::Value,
+    pub method: String,
+    pub params: T,
 }
